Don't run dehydrated unless necessary
Reason for this is, that dehydrated queries the letsencrypt servers on
every run, even when the certificate is still valid.
If the letsencrypt server is down for maintenance or the request fails
due to another reason (happens regularly for multiple servers performing
the query every 5 minutes = 288 times per day), error messages are
generated, which is very annoying for the sysadmin.